+++ backend/app/Http/Controllers/PlacementRequest/FinalizePlacementRequestController.php
@@ -121,7 +121,7 @@ class FinalizePlacementRequestController extends Controller
                     NotificationType::PLACEMENT_ENDED->value,
                     [
                         'message' => 'The placement for '.$placementRequest->pet->name.' has ended. Thank you for your help!',
-                        'link' => '/pets/'.$placementRequest->pet_id.'/public',
+                        'link' => '/pets/'.$placementRequest->pet_id.'/view',
                         'pet_name' => $placementRequest->pet->name,
                         'pet_id' => $placementRequest->pet_id,
                     ]